+++ Assets/Scripts/GameManager.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public int currentGold = 0;
+    public float currentGold = 0f; // Ubah ke float sesuai dengan goldPerRound di BuildingGoldMine
 
-    public void AddGold(int amount)
+    public void AddGold(float amount) // Ubah parameter ke float
     {
         currentGold += amount;
         Debug.Log("Gold bertambah: " + amount + ", total: " + currentGold);
